pyproject.toml
[tool.poetry]
name = "visivo"
version = "0.1.0"
description = ""
authors = ["Visivo People"]

[tool.poetry.scripts]
visivo = "visivo.command_line:visivo"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^1.10.5"
requests = "^2.28.2"
Jinja2 = "^3.1.2"
PyYAML = "^6.0"
sql-formatter = "^0.6.2"
SQLAlchemy = "^2.0.4"
click = "^8.1.3"
pandas = "^1.5.3"
Flask = "^2.2.3"
livereload = "^2.6.3"
psycopg2-binary = "^2.9.5"

[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
factory-boy = "^3.2.1"
mkdocs = "^1.4.2"
mkdocs-click = "^0.8.0"
mkdocs-encryptcontent-plugin = "^2.4.5"
mkdocs-material = "^9.1.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"